Debugimplementasjon av servo og HCSR04-sensor

+++ ZUMO/ultrasonic.cpp
@@ -4,13 +4,14 @@
 #define echoPin 17 // attach pin D2 Arduino to pin Echo of HC-SR04
 #define trigPin 14 //attach pin D3 Arduino to pin Trig of HC-SR04
 
+
 long duration; // variable for the duration of sound wave travel
-int distance; // variable for the distance measurement
+uint8_t distance; // variable for the distance measurement
 unsigned long period_1 = 2;
 unsigned long period_2 = 10;
 unsigned long timenow = 0;
 
-int ultrasonic() {
+uint8_t ultrasonic() {
   // Clears the trigPin condition
   while (micros() > timenow + period_1) {
     timenow += micros();
